Cleaning up WorldGen code
